requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.5.7
Flask-WTF==0.12
GitPython==1.0.2
Jinja2==2.8
MarkupSafe==0.23
Pygments==2.1.1
WTForms==2.1
Werkzeug==0.11.4
Whoosh==2.7.2
argcomplete==1.1.0
argh==0.26.1
blessings==1.6
dominate==2.1.17
gitdb==0.6.4
itsdangerous==0.24
pyperclip==1.5.26
python-magic==0.4.10
smmap==0.9.0
visitor==0.1.2
